docs(react-motions): restructure docs
