+++ packages/react-components/react-motions-preview/stories/CreateMotionComponent/CreateMotionComponentDescription.md
@@ -1,0 +1,7 @@
+`createMotionComponent()` is a factory function that creates a custom motion component powered by Web Animations API.
+
+<!-- Don't allow prettier to collapse code block into single line -->
+<!-- prettier-ignore -->
+> **⚠️ Preview packages are considered unstable:**
+> - Features and APIs may change before final release
+> - Please contact us if you intend to use this in your product
