+++ packages/react-components/react-motions-preview/stories/CreateMotionComponent/TokensUsage.stories.md
@@ -1,0 +1,1 @@
+Fluent UI provides a set of design tokens, via `tokens` object, for example `tokens.colorNeutralForeground1`. You can use them to create your own motions.
