+++ packages/react-components/react-motions-preview/stories/CreatePresenceComponent/CreatePresenceComponent.stories.tsx
@@ -1,5 +1,5 @@
-import { makeStyles, shorthands, tokens, Label, Slider, useId, Checkbox } from '@fluentui/react-components';
-import { createPresenceComponent, motionTokens } from '@fluentui/react-motions-preview';
+import { makeStyles, shorthands, tokens, Checkbox } from '@fluentui/react-components';
+import { createPresenceComponent } from '@fluentui/react-motions-preview';
 import type { MotionImperativeRef } from '@fluentui/react-motions-preview';
 import * as React from 'react';
 
@@ -43,61 +43,41 @@ const useClasses = makeStyles({
   },
 });
 
-const Fade = createPresenceComponent({
+const DropIn = createPresenceComponent({
   enter: {
-    keyframes: [{ opacity: 0 }, { opacity: 1 }],
-    duration: motionTokens.durationSlow,
+    keyframes: [
+      { transform: 'rotate(-30deg) translateY(-100%)', opacity: 0 },
+      { transform: 'rotate(0deg) translateY(0%)', opacity: 1 },
+    ],
+    duration: 2000,
   },
   exit: {
     keyframes: [{ opacity: 1 }, { opacity: 0 }],
-    duration: motionTokens.durationSlow,
+    duration: 1000,
   },
 });
 
 export const CreatePresenceComponent = () => {
   const classes = useClasses();
-  const sliderId = useId();
 
   const motionRef = React.useRef<MotionImperativeRef>();
-
-  const [playbackRate, setPlaybackRate] = React.useState<number>(30);
   const [visible, setVisible] = React.useState<boolean>(false);
-
-  // Heads up!
-  // This is optional and is intended solely to slow down the animations, making motions more visible in the examples.
-  React.useEffect(() => {
-    motionRef.current?.setPlaybackRate(playbackRate / 100);
-  }, [playbackRate, visible]);
 
   return (
     <>
       <div className={classes.container}>
         <div className={classes.card}>
-          <Fade imperativeRef={motionRef} visible={visible}>
+          <DropIn imperativeRef={motionRef} visible={visible}>
             <div className={classes.item} />
-          </Fade>
+          </DropIn>
 
-          <code className={classes.description}>fadeSlow</code>
+          <code className={classes.description}>Drop in</code>
         </div>
       </div>
 
       <div className={classes.controls}>
         <div>
           <Checkbox label={<code>visible</code>} checked={visible} onChange={() => setVisible(v => !v)} />
-        </div>
-        <div>
-          <Label htmlFor={sliderId}>
-            <code>playbackRate</code>: {playbackRate}%
-          </Label>
-          <Slider
-            aria-valuetext={`Value is ${playbackRate}%`}
-            value={playbackRate}
-            onChange={(ev, data) => setPlaybackRate(data.value)}
-            min={0}
-            id={sliderId}
-            max={100}
-            step={10}
-          />
         </div>
       </div>
     </>
