+++ packages/react-components/react-motions-preview/stories/CreatePresenceComponent/CreatePresenceComponentDefault.stories.tsx
@@ -1,0 +1,77 @@
+import { makeStyles, shorthands, tokens, Checkbox } from '@fluentui/react-components';
+import { createPresenceComponent, motionTokens } from '@fluentui/react-motions-preview';
+import * as React from 'react';
+
+const useClasses = makeStyles({
+  container: {
+    display: 'grid',
+    gridTemplateColumns: '1fr',
+    ...shorthands.gap('10px'),
+  },
+  card: {
+    display: 'flex',
+    flexDirection: 'column',
+
+    ...shorthands.border('3px', 'solid', tokens.colorNeutralForeground3),
+    ...shorthands.borderRadius(tokens.borderRadiusMedium),
+    ...shorthands.padding('10px'),
+
+    alignItems: 'center',
+  },
+  item: {
+    backgroundColor: tokens.colorBrandBackground,
+    ...shorthands.borderRadius('50%'),
+
+    width: '100px',
+    height: '100px',
+  },
+  description: {
+    ...shorthands.margin('5px'),
+  },
+  controls: {
+    display: 'flex',
+    flexDirection: 'column',
+
+    marginTop: '20px',
+
+    ...shorthands.border('3px', 'solid', tokens.colorNeutralForeground3),
+    ...shorthands.borderRadius(tokens.borderRadiusMedium),
+    ...shorthands.padding('10px'),
+  },
+});
+
+const Fade = createPresenceComponent({
+  enter: {
+    keyframes: [{ opacity: 0 }, { opacity: 1 }],
+    duration: motionTokens.durationSlow,
+  },
+  exit: {
+    keyframes: [{ opacity: 1 }, { opacity: 0 }],
+    duration: motionTokens.durationSlow,
+  },
+});
+
+export const CreatePresenceComponentDefault = () => {
+  const classes = useClasses();
+  const [visible, setVisible] = React.useState<boolean>(false);
+
+  return (
+    <>
+      <div className={classes.container}>
+        <div className={classes.card}>
+          <Fade visible={visible}>
+            <div className={classes.item} />
+          </Fade>
+
+          <code className={classes.description}>fadeSlow</code>
+        </div>
+      </div>
+
+      <div className={classes.controls}>
+        <div>
+          <Checkbox label={<code>visible</code>} checked={visible} onChange={() => setVisible(v => !v)} />
+        </div>
+      </div>
+    </>
+  );
+};
