+++ packages/react-components/react-motions-preview/stories/CreatePresenceComponent/CreatePresenceComponentDescription.md
@@ -1,0 +1,7 @@
+`createPresenceComponent()` is a factory function that creates a React component based on the provided presence definition. This component can be used to animate any element and intended to have a state via the `visible` prop.
+
+<!-- Don't allow prettier to collapse code block into single line -->
+<!-- prettier-ignore -->
+> **⚠️ Preview packages are considered unstable:**
+> - Features and APIs may change before final release
+> - Please contact us if you intend to use this in your product
