--- packages/react-components/react-motions-preview/stories/Motion/CreateMotionComponent.stories.md
@@ -1,1 +1,0 @@
-`createMotionComponent()` is a factory function that creates a React component based on the provided atom definition. This component can be used to animate any element.
