--- packages/react-components/react-motions-preview/stories/Motion/TokensUsage.stories.md
@@ -1,1 +1,0 @@
-As Fluent UI provides a set of design tokens, via `tokens` object, for example `tokens.colorNeutralForeground1`. You can use them to create your own motions.
